+++ app/src/main/java/xyz/skether/radiline/domain/PlaylistManager.kt
@@ -5,7 +5,7 @@ import javax.inject.Inject
 
 class PlaylistManager @Inject constructor(private val api: ShoutcastApi) {
 
-    suspend fun getStationTrack(stationId: Int): Track? {
+    suspend fun getStationTrack(stationId: Long): Track? {
         // todo load tuneIn from DB
         val trackListResp = api.getPlaylist(stationId, "/sbin/tunein-station.xspf")
         val trackResp = trackListResp.trackList.firstOrNull()
